[SB][91144750] add showInfoWindow event trigger to markers
Allows anything to trigger an event to popup an info window
over a marker. Kicks off the same code as clicking a marker
(which makes an ajax request for the pin content).
In actual use it's way too slow, especially if kicked off
for each item in the list as the user sweeps their mouse over
several list results.

To make it performant enough, I think we need to load the
info window content (or the data to render it client
side via handlebars) when we fetch the pins data.
Will explore that idea in a new branch.