[SB][105361098] Enable cluster avg centering
Since we weren't setting this, it was defaulting to false,
so each cluster was positioned based on the first marker added
to it.
This became more visibly obvious with the new decluster animation...
I noticed that instead of spreading out from a central point, one
seemingly random marker would stay put and the other markers would
animate away from it.